# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "properties": {
          "code": {
            "type": "string"
          },
          "details": true,
          "error": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "details",
          "error"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
...
      "PluginInstallRequest": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Request body for POST /api/plugins/install",
        "properties": {
          "scope": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "source"
        ],
        "title": "PluginInstallRequest",
        "type": "object"
      },
      "PluginInstallResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Response body for POST /api/plugins/install",
        "properties": {
          "fresh_install": {
            "type": "boolean"
          },
          "install_path": {
            "type": "string"
          },
          "plugin": true
        },
        "required": [
          "fresh_install",
          "install_path",
          "plugin"
        ],
        "title": "PluginInstallResponse",
        "type": "object"
      },
...
      "PluginsListResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Response to GET /api/plugins",
        "properties": {
          "diagnostics": {
            "items": true,
            "type": "array"
          },
          "plugins": {
            "items": true,
            "type": "array"
          }
        },
        "required": [
          "diagnostics",
          "plugins"
        ],
        "title": "PluginsListResponse",
        "type": "object"
      },
      "PluginsMarketplaceResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "description": "Response to GET /api/plugins/marketplace",
        "properties": {
          "plugins": {
            "items": true,
            "type": "array"
          },
          "sources": {
            "items": true,
            "type": "array"
          }
        },
        "required": [
          "plugins",
          "sources"
        ],
        "title": "PluginsMarketplaceResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "allthecodes Web API",
    "version": "v1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/plugins": {
      "get": {
        "operationId": "PluginsList",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginsListResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      }
    },
    "/api/plugins/install": {
      "post": {
        "operationId": "PluginsInstall",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PluginInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginInstallResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      }
    },
    "/api/plugins/marketplace": {
      "get": {
        "operationId": "PluginsMarketplace",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginsMarketplaceResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "API error"
          }
        },
        "x-allthecodes-serialization": "concurrent"
      }
    },
...
  }
}